Moved SpawnCost namespace to Properties
