+++ ExtraEnemyCustomization/Configs/Customizations/PropertyCustomConfig.cs
@@ -1,4 +1,4 @@
-﻿using EECustom.Customizations.SpawnCost;
+﻿using EECustom.Customizations.Properties;
 using System;
 using System.Collections.Generic;
 using System.Text;
