+++ ExtraEnemyCustomization/Configs/Customizations/SpawnCostCustomConfig.cs
@@ -1,5 +1,5 @@
 ﻿using EECustom.Customizations;
-using EECustom.Customizations.SpawnCost;
+using EECustom.Customizations.Properties;
 using System;
 using System.Collections.Generic;
